Backend/requirements.txt
flask==2.0.1
flask-cors==3.0.10
pymongo==4.1.1
PyJWT==2.3.0
werkzeug==2.0.2
python-dotenv==0.19.2
bson==0.5.10
